Reorganized database initialization script
The sample data insertion for the 'accounts' table has been moved in the script. Previously, it was inserted just after the creation of the 'accounts' table. However, having the sample data scattered around was confusing and disorganized. To make the script more efficient and readable, all of the sample data insertions were grouped together at the end of the script. This change does not affect the functionality of the code, but improves its clarity, organization, and readability.